requirements.txt
pygame~=2.5.2
pyserial~=3.5